Add files via upload
Dear Sir, sorry this is so late. I spent a good amount of time on this. I think that I understand the theory quite well and I understand how to answer the question, but I always seem to make mistakes in the code or the code does not reflect what I wanted to do.